reuse reqwest::client, but a bit ugly
